Improve GUID syntax highlighting
The existing GUID syntax matching for UEFI filetypes was not
sufficiently precise. The C-style GUID match was a broad region,
and the standard GUID match was generically named.

Replacing the decCFormatGUID region with two more specific match
patterns: cGUID for standard C-style GUIDs and cNullGuid for
the NULL GUID representation.